+++ CHANGELOG.md
@@ -5,12 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `unreleased`
+
+### Added
+
+- Add odpy convert `input_table` input through config file, which gives the
+ability to list multiple file glob expression and associated metadata.
+
 ## `0.5.2` - 2024-06-22
 
 ### Fixed
 
 - Docs build
-
 
 ## `0.5.1` - 2024-06-20